Expressions & Operators: Better expression parsing to support binary expressions & user-defined operators.
The Dot `.` token is no longer reserved as a special `TokenType`, allowing for user-defined operators to use
it as a valid symbol.

Redefined the syntax for method calls. Valid calls now look like:
	`call RECEIVER MESSAGE ARGS_TUPLE` e.g. `call Int add (1, 2)`

NOTE: Although this new call syntax is awkward, once quoting is implemented, it should be possible
to define an infix Dot operator within the Standard Library to reintroduce the familiar dot-syntax, e.g:
	`Int.add(1, 2)`