Types: Model Attributes & Type Effects in the type system, rather than pass around INodes for deferred evaluation.
The previous evaluation model for Attributes was to pass around the original body node, and delay type checking until it was
actually invoked. This allowed type errors to slip through if the Attribute was never used, and also made type checking of
Type Lambdas more difficult to reason about in general.